WebAssembly doesn't have i64 so overflows on negative values. Use isize instead.

+++ src/ast.rs
@@ -26,7 +26,7 @@ pub enum Expr {
         right: Box<Expr>,
     },
     Int {
-        value: i64,
+        value: isize,
     },
     Float {
         value: f64,
